Add files via upload
This is the code (beta) for the tutorial coded in R and greta. 
It is work in progress. The goal of the code is to show how to (1) simulate data under the model, (2) fit the model and (3) analyze model fit. The current work in progress consists of getting the estimation of a known variance hierarchy (i.e., sample, host and phylo) right.